app/requirements.txt
# User Service Python Dependencies
# Core packages for User service in MSA environment

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Swagger-UI==4.11.1
Werkzeug==2.3.7
python-dotenv==1.0.0
boto3==1.34.0
python-jose==3.3.0
requests==2.31.0
